Feat: Make package registry downward compatible (#74)
The release of the Full Package History in product-core needs to be non breaking (or backward compatible) so that products using the already existing version of the PackageRegistry can use the new version without changes.

For this purpose this PR changes the following:

* Switch toml dependency back to be static.
    To prevent existing projects from having build issues, the static toml depency needs to be maintained.
    As soon as all products have migrated to use the MoveHistoryManager and the legacy PackageRegistry functions
    (marked as deprecated now) have been removed, we can reactivate the toml dependency being controlled by
    the move-history-manager feature flag.

* Make Package Registry downward compatible.
    The legacy versions of the functions insert_env and from_move_lock_content have been restored and the new
    version of insert_env has been renamed to insert_env_history to be distinguishable.

Products using legacy PackageRegistry functions that have been marked as deprecated now need to add the #[allow(deprecated)] macro everywhere where those deprecated functions and types are used in their code.